+++ _test/tests/inc/parser/parser_quotes.test.php
@@ -181,13 +181,8 @@ class TestOfDoku_Parser_Quotes extends TestOfDoku_Parser {
 
     function testDoubleQuoteClosing() {
         $raw = 'Foo hello" Bar';
-        $this->P->addMode('quotes',new Quotes());
-
-        /** @noinspection PhpUnhandledExceptionInspection */
-        $status = $this->getInaccessibleProperty($this->H, 'status');
-        $status['doublequote'] = 1;
-        /** @noinspection PhpUnhandledExceptionInspection */
-        $this->setInaccessibleProperty($this->H, 'status', $status);
+        $this->P->addMode('quotes', new Quotes());
+        $this->H->setStatus('doublequote', 1);
         $this->P->parse($raw);
 
         $calls = array (
@@ -205,12 +200,8 @@ class TestOfDoku_Parser_Quotes extends TestOfDoku_Parser {
 
     function testDoubleQuoteClosingSpecial() {
         $raw = 'Foo hello") Bar';
-        $this->P->addMode('quotes',new Quotes());
-        /** @noinspection PhpUnhandledExceptionInspection */
-        $status = $this->getInaccessibleProperty($this->H, 'status');
-        $status['doublequote'] = 1;
-        /** @noinspection PhpUnhandledExceptionInspection */
-        $this->setInaccessibleProperty($this->H, 'status', $status);
+        $this->P->addMode('quotes', new Quotes());
+        $this->H->setStatus('doublequote', 1);
 
         $this->P->parse($raw);
 
@@ -228,12 +219,7 @@ class TestOfDoku_Parser_Quotes extends TestOfDoku_Parser {
     }
     function testDoubleQuoteClosingSpecial2() {
         $raw = 'Foo hello") Bar';
-        $this->P->addMode('quotes',new Quotes());
-        /** @noinspection PhpUnhandledExceptionInspection */
-        $status = $this->getInaccessibleProperty($this->H, 'status');
-        $status['doublequote'] = 0;
-        /** @noinspection PhpUnhandledExceptionInspection */
-        $this->setInaccessibleProperty($this->H, 'status', $status);
+        $this->P->addMode('quotes', new Quotes());
 
         $this->P->parse($raw);
