+++ inc/Parsing/Lexer/Lexer.php
@@ -10,6 +10,8 @@
 
 namespace dokuwiki\Parsing\Lexer;
 
+use dokuwiki\Parsing\Handler;
+
 /**
  * Accepts text and breaks it into tokens.
  *
@@ -25,7 +27,7 @@ class Lexer
 
     /** @var ParallelRegex[] */
     protected $regexes = [];
-    /** @var \Doku_Handler */
+    /** @var Handler */
     protected $handler;
     /** @var StateStack */
     protected $modeStack;
@@ -37,7 +39,7 @@ class Lexer
     /**
      * Sets up the lexer in case insensitive matching by default.
      *
-     * @param \Doku_Handler $handler  Handling strategy by reference.
+     * @param Handler $handler  Handling strategy by reference.
      * @param string $start            Starting handler.
      * @param boolean $case            True for case sensitive.
      */
@@ -246,9 +248,10 @@ class Lexer
     }
 
     /**
-     * Calls the parser method named after the current mode.
+     * Dispatches a token to the handler.
      *
-     * Empty content will be ignored. The lexer has a parser handler for each mode in the lexer.
+     * Resolves mode name aliases (e.g. unformattedalt → unformatted) and
+     * delegates all dispatch logic to Handler::handleToken().
      *
      * @param string $content Text parsed.
      * @param boolean $is_match Token is recognised rather
@@ -262,19 +265,10 @@ class Lexer
         if (($content === "") || ($content === false)) {
             return true;
         }
-        $handler = $this->modeStack->getCurrent();
-        if (isset($this->mode_handlers[$handler])) {
-            $handler = $this->mode_handlers[$handler];
-        }
+        $originalName = $this->modeStack->getCurrent();
+        $modeName = $this->mode_handlers[$originalName] ?? $originalName;
 
-        // modes starting with plugin_ are all handled by the same
-        // handler but with an additional parameter
-        if (str_starts_with($handler, 'plugin_')) {
-            [$handler, $plugin] = sexplode('_', $handler, 2, '');
-            return $this->handler->$handler($content, $is_match, $pos, $plugin);
-        }
-
-        return $this->handler->$handler($content, $is_match, $pos);
+        return $this->handler->handleToken($modeName, $content, $is_match, $pos, $originalName);
     }
 
     /**
