+++ inc/Parsing/ParserMode/Listblock.php
@@ -2,6 +2,8 @@
 
 namespace dokuwiki\Parsing\ParserMode;
 
+use dokuwiki\Parsing\Handler;
+use dokuwiki\Parsing\Handler\Lists;
 use dokuwiki\Parsing\ModeRegistry;
 
 class Listblock extends AbstractMode
@@ -17,6 +19,12 @@ class Listblock extends AbstractMode
             ModeRegistry::CATEGORY_DISABLED,
             ModeRegistry::CATEGORY_PROTECTED,
         ]);
+    }
+
+    /** @inheritdoc */
+    public function getSort()
+    {
+        return 10;
     }
 
     /** @inheritdoc */
@@ -44,8 +52,26 @@ class Listblock extends AbstractMode
     }
 
     /** @inheritdoc */
-    public function getSort()
+    public function handle($match, $state, $pos, Handler $handler)
     {
-        return 10;
+        switch ($state) {
+            case DOKU_LEXER_ENTER:
+                $handler->setCallWriter(new Lists($handler->getCallWriter()));
+                $handler->addCall('list_open', [$match], $pos);
+                break;
+            case DOKU_LEXER_EXIT:
+                $handler->addCall('list_close', [], $pos);
+                /** @var Lists $reWriter */
+                $reWriter = $handler->getCallWriter();
+                $handler->setCallWriter($reWriter->process());
+                break;
+            case DOKU_LEXER_MATCHED:
+                $handler->addCall('list_item', [$match], $pos);
+                break;
+            case DOKU_LEXER_UNMATCHED:
+                $handler->addCall('cdata', [$match], $pos);
+                break;
+        }
+        return true;
     }
 }
