+++ inc/Parsing/ParserMode/ModeInterface.php
@@ -2,6 +2,8 @@
 
 namespace dokuwiki\Parsing\ParserMode;
 
+use dokuwiki\Parsing\Handler;
+
 /**
  * Defines a mode (syntax component) in the Parser
  */
@@ -43,4 +45,15 @@ interface ModeInterface
      * @return bool
      */
     public function accepts($mode);
+
+    /**
+     * Handle a matched token from the lexer.
+     *
+     * @param string $match The matched text
+     * @param int $state The lexer state (DOKU_LEXER_ENTER, _EXIT, _MATCHED, etc.)
+     * @param int $pos Byte position in the source
+     * @param Handler $handler The handler (for addCall, status, etc.)
+     * @return bool
+     */
+    public function handle($match, $state, $pos, Handler $handler);
 }
