+++ inc/Parsing/ParserMode/Multiplyentity.php
@@ -2,11 +2,19 @@
 
 namespace dokuwiki\Parsing\ParserMode;
 
+use dokuwiki\Parsing\Handler;
+
 /**
  * Implements the 640x480 replacement
  */
 class Multiplyentity extends AbstractMode
 {
+    /** @inheritdoc */
+    public function getSort()
+    {
+        return 270;
+    }
+
     /** @inheritdoc */
     public function connectTo($mode)
     {
@@ -19,8 +27,10 @@ class Multiplyentity extends AbstractMode
     }
 
     /** @inheritdoc */
-    public function getSort()
+    public function handle($match, $state, $pos, Handler $handler)
     {
-        return 270;
+        preg_match_all('/\d+/', $match, $matches);
+        $handler->addCall('multiplyentity', [$matches[0][0], $matches[0][1]], $pos);
+        return true;
     }
 }
