+++ inc/Parsing/ParserMode/Nocache.php
@@ -2,8 +2,16 @@
 
 namespace dokuwiki\Parsing\ParserMode;
 
+use dokuwiki\Parsing\Handler;
+
 class Nocache extends AbstractMode
 {
+    /** @inheritdoc */
+    public function getSort()
+    {
+        return 40;
+    }
+
     /** @inheritdoc */
     public function connectTo($mode)
     {
@@ -11,8 +19,9 @@ class Nocache extends AbstractMode
     }
 
     /** @inheritdoc */
-    public function getSort()
+    public function handle($match, $state, $pos, Handler $handler)
     {
-        return 40;
+        $handler->addCall('nocache', [], $pos);
+        return true;
     }
 }
