+++ inc/Parsing/ParserMode/Preformatted.php
@@ -2,10 +2,18 @@
 
 namespace dokuwiki\Parsing\ParserMode;
 
+use dokuwiki\Parsing\Handler;
+use dokuwiki\Parsing\Handler\Preformatted as PreformattedHandler;
 use dokuwiki\Parsing\ModeRegistry;
 
 class Preformatted extends AbstractMode
 {
+    /** @inheritdoc */
+    public function getSort()
+    {
+        return 20;
+    }
+
     /** @inheritdoc */
     public function connectTo($mode)
     {
@@ -27,8 +35,27 @@ class Preformatted extends AbstractMode
     }
 
     /** @inheritdoc */
-    public function getSort()
+    public function handle($match, $state, $pos, Handler $handler)
     {
-        return 20;
+        switch ($state) {
+            case DOKU_LEXER_ENTER:
+                $handler->setCallWriter(new PreformattedHandler($handler->getCallWriter()));
+                $handler->addCall('preformatted_start', [], $pos);
+                break;
+            case DOKU_LEXER_EXIT:
+                $handler->addCall('preformatted_end', [], $pos);
+                /** @var PreformattedHandler $reWriter */
+                $reWriter = $handler->getCallWriter();
+                $handler->setCallWriter($reWriter->process());
+                break;
+            case DOKU_LEXER_MATCHED:
+                $handler->addCall('preformatted_newline', [], $pos);
+                break;
+            case DOKU_LEXER_UNMATCHED:
+                $handler->addCall('preformatted_content', [$match], $pos);
+                break;
+        }
+
+        return true;
     }
 }
