+++ inc/Parsing/ParserMode/Quote.php
@@ -2,6 +2,8 @@
 
 namespace dokuwiki\Parsing\ParserMode;
 
+use dokuwiki\Parsing\Handler;
+use dokuwiki\Parsing\Handler\Quote as QuoteHandler;
 use dokuwiki\Parsing\ModeRegistry;
 
 class Quote extends AbstractMode
@@ -20,6 +22,12 @@ class Quote extends AbstractMode
     }
 
     /** @inheritdoc */
+    public function getSort()
+    {
+        return 220;
+    }
+
+    /** @inheritdoc */
     public function connectTo($mode)
     {
         $this->Lexer->addEntryPattern('\n>{1,}', $mode, 'quote');
@@ -33,8 +41,30 @@ class Quote extends AbstractMode
     }
 
     /** @inheritdoc */
-    public function getSort()
+    public function handle($match, $state, $pos, Handler $handler)
     {
-        return 220;
+        switch ($state) {
+            case DOKU_LEXER_ENTER:
+                $handler->setCallWriter(new QuoteHandler($handler->getCallWriter()));
+                $handler->addCall('quote_start', [$match], $pos);
+                break;
+
+            case DOKU_LEXER_EXIT:
+                $handler->addCall('quote_end', [], $pos);
+                /** @var QuoteHandler $reWriter */
+                $reWriter = $handler->getCallWriter();
+                $handler->setCallWriter($reWriter->process());
+                break;
+
+            case DOKU_LEXER_MATCHED:
+                $handler->addCall('quote_newline', [$match], $pos);
+                break;
+
+            case DOKU_LEXER_UNMATCHED:
+                $handler->addCall('cdata', [$match], $pos);
+                break;
+        }
+
+        return true;
     }
 }
