+++ inc/Parsing/ParserMode/Smiley.php
@@ -2,6 +2,7 @@
 
 namespace dokuwiki\Parsing\ParserMode;
 
+use dokuwiki\Parsing\Handler;
 use dokuwiki\Parsing\Lexer\Lexer;
 
 class Smiley extends AbstractMode
@@ -16,6 +17,12 @@ class Smiley extends AbstractMode
     public function __construct($smileys)
     {
         $this->smileys = $smileys;
+    }
+
+    /** @inheritdoc */
+    public function getSort()
+    {
+        return 230;
     }
 
     /** @inheritdoc */
@@ -41,8 +48,9 @@ class Smiley extends AbstractMode
     }
 
     /** @inheritdoc */
-    public function getSort()
+    public function handle($match, $state, $pos, Handler $handler)
     {
-        return 230;
+        $handler->addCall('smiley', [$match], $pos);
+        return true;
     }
 }
