+++ inc/Parsing/ParserMode/Table.php
@@ -2,6 +2,8 @@
 
 namespace dokuwiki\Parsing\ParserMode;
 
+use dokuwiki\Parsing\Handler;
+use dokuwiki\Parsing\Handler\Table as TableHandler;
 use dokuwiki\Parsing\ModeRegistry;
 
 class Table extends AbstractMode
@@ -17,6 +19,12 @@ class Table extends AbstractMode
             ModeRegistry::CATEGORY_DISABLED,
             ModeRegistry::CATEGORY_PROTECTED,
         ]);
+    }
+
+    /** @inheritdoc */
+    public function getSort()
+    {
+        return 60;
     }
 
     /** @inheritdoc */
@@ -45,8 +53,55 @@ class Table extends AbstractMode
     }
 
     /** @inheritdoc */
-    public function getSort()
+    public function handle($match, $state, $pos, Handler $handler)
     {
-        return 60;
+        switch ($state) {
+            case DOKU_LEXER_ENTER:
+                $handler->setCallWriter(new TableHandler($handler->getCallWriter()));
+
+                $handler->addCall('table_start', [$pos + 1], $pos);
+                if (trim($match) == '^') {
+                    $handler->addCall('tableheader', [], $pos);
+                } else {
+                    $handler->addCall('tablecell', [], $pos);
+                }
+                break;
+
+            case DOKU_LEXER_EXIT:
+                $handler->addCall('table_end', [$pos], $pos);
+                /** @var TableHandler $reWriter */
+                $reWriter = $handler->getCallWriter();
+                $handler->setCallWriter($reWriter->process());
+                break;
+
+            case DOKU_LEXER_UNMATCHED:
+                if (trim($match) != '') {
+                    $handler->addCall('cdata', [$match], $pos);
+                }
+                break;
+
+            case DOKU_LEXER_MATCHED:
+                if ($match == ' ') {
+                    $handler->addCall('cdata', [$match], $pos);
+                } elseif (preg_match('/:::/', $match)) {
+                    $handler->addCall('rowspan', [$match], $pos);
+                } elseif (preg_match('/\t+/', $match)) {
+                    $handler->addCall('table_align', [$match], $pos);
+                } elseif (preg_match('/ {2,}/', $match)) {
+                    $handler->addCall('table_align', [$match], $pos);
+                } elseif ($match == "\n|") {
+                    $handler->addCall('table_row', [], $pos);
+                    $handler->addCall('tablecell', [], $pos);
+                } elseif ($match == "\n^") {
+                    $handler->addCall('table_row', [], $pos);
+                    $handler->addCall('tableheader', [], $pos);
+                } elseif ($match == '|') {
+                    $handler->addCall('tablecell', [], $pos);
+                } elseif ($match == '^') {
+                    $handler->addCall('tableheader', [], $pos);
+                }
+                break;
+        }
+        return true;
     }
 }
